static_cdn/static_cdn/requirements.txt
astroid==1.4.9
colorama==0.3.7
Django==1.10.2
django-ajax==0.2.0
django-ajax-selects==1.5.1
django-appconf==1.0.2
django-bootstrap-form==3.2.1
django-excel==0.0.6
django-mailer==1.2.2
django-markdown-deux==1.0.5
django-notification==1.3.3
django-pagedown==0.1.1
django-pagination==1.0.7
django-postman==3.4.0
django-user-accounts==1.3.1
et-xmlfile==1.0.1
gunicorn==19.6.0
isort==4.2.5
jdcal==1.3
lazy-object-proxy==1.2.2
lockfile==0.12.2
markdown2==2.3.1
mccabe==0.6.1
odfpy==1.3.4
openpyxl==2.4.5
Pillow==3.4.1
pinax-notifications==3.0.1
pyexcel==0.4.5
pyexcel-io==0.3.3
pyexcel-ods==0.3.1
pyexcel-webio==0.0.11
pyexcel-xls==0.3.1
pyexcel-xlsx==0.3.0
pyflakes==1.5.0
pylint==1.6.5
pytz==2016.7
six==1.10.0
texttable==0.8.8
wrapt==1.10.10
xlrd==1.0.0
xlwt==1.2.0
